requirements.txt
appnope==0.1.0
decorator==4.0.6
Django==1.9.1
django-ckeditor==5.0.3
django-cors-headers==1.1.0
django-filter==0.12.0
django-model-utils==2.4
django-rest-auth==0.6.0
django-taggit==0.18.0
django-wysiwyg==0.7.1
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
gnureadline==6.3.3
gunicorn==19.4.5
html5lib==0.9999999
ipdb==0.8.1
ipython==4.0.3
ipython-genutils==0.1.0
Markdown==2.6.5
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5
PyJWT==1.4.0
requests==2.9.1
requests-oauthlib==0.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
wheel==0.24.0
